feat: 调整字段的类型

See merge request ckb-explorer/ckb-explorer-web!37

+++ src/main/java/com/ckb/explorer/domain/dto/TransactionDto.java
@@ -28,7 +28,7 @@ public class TransactionDto {
 
   private Long blockTimestamp;
 
-  private Integer cycles;
+  private Long cycles;
 
   private Long bytes;
